Updated markEntitySentiment for performance
Code now takes JSON response data returned from retrieveEntitySentiment, creates an array of the data for the "Entity Sentiment Data" tab, and pastes the array as a range in the sheet. Previously, these results were pasted row by row.

Also removed the code from flagging of the first row to start processing; will now start checking for new rows to process from to the top of the sheet each time markEntitySentiment is run.